compile

+++ lib/FCIdump.cpp
@@ -253,14 +253,14 @@ FCIdump::integralType FCIdump::nextIntegral(int& i, int& j, int& k, int& l, doub
   return result;
 }
 
-FCIdump::integralType FCIdump::nextIntegral(unsigned int& symi,
-                                            size_t& i,
-                                            unsigned int& symj,
-                                            size_t& j,
-                                            unsigned int& symk,
-                                            size_t& k,
-                                            unsigned int& syml,
-                                            size_t& l,
+FCIdump::integralType FCIdump::nextIntegral(int& symi,
+                                            int& i,
+                                            int& symj,
+                                            int& j,
+                                            int& symk,
+                                            int& k,
+                                            int& syml,
+                                            int& l,
                                             double& value) const {
   int ii, jj, kk, ll;
   auto result = nextIntegral(ii, jj, kk, ll, value);
